requirements.txt
flask
python-dotenv
flask-sqlalchemy
psycopg2-binary
flask-bcrypt
flask-login
flask-wtf
email_validator
flask-cors